Move cart related functions to their own file
